Refactored Header, SortControl, EditableRecipeDetails, ImageGallery; updated imports and code organization
