+++ app/components/Header.js
@@ -5,11 +5,11 @@ import { useState, useEffect, Suspense } from "react";
 import { FaUser } from "react-icons/fa";
 import { useRouter } from "next/navigation";
 import CategoryList from "./ui/CategoryList.js";
-import { FilterButton } from "./ui/FilterButton";
+import { FilterButton } from "./filter-sort/FilterButton.js";
 import ThemeButton from "./ui/ThemeButton";
 import RecipeSearchBar from "./ui/searchBar.js";
 import UserModal from "./UserModal.js";
-import { FilterModal } from "./ui/FilterButton";
+import { FilterModal } from "./filter-sort/FilterButton.js";
 import { ShoppingListProvider } from "../context/ShoppingListContext.js";
 import ShoppingBagHeader from "./ShoppingBagHeader.js";
 import Loading from "../loading.js";
