+++ app/components/filter-sort/FilterButton.js
@@ -6,6 +6,14 @@ import { fetchRecipes } from "../../api";
 import IngDisplay from "../IngredientList";
 import StepsDropdown from "../StepsDropdown";
 
+/**
+ * Modal component for filtering recipes with various search criteria.
+ * 
+ * @component
+ * @param {Object} props - Component props
+ * @param {Function} props.onClose - Callback function to close the filter modal
+ * @returns {JSX.Element} Rendered filter modal
+ */
 export const FilterModal = ({ onClose }) => {
     const [cookTime, setCookTime] = useState(0);
     const [selectedTags, setSelectedTags] = useState([]);
@@ -14,6 +22,12 @@ export const FilterModal = ({ onClose }) => {
     const searchParams = useSearchParams();
     const router = useRouter();
 
+    /**
+     * Converts a numeric time value to a human-readable range label.
+     * 
+     * @param {number} value - The time value to convert
+     * @returns {string} Formatted time range label
+     */
     const getRangeLabel = (value) => {
         if (value === 0) return "All";
         if (value <= 15) return "5-15 min";
@@ -22,6 +36,12 @@ export const FilterModal = ({ onClose }) => {
         return "45-60 min";
     };
 
+    /**
+     * Snaps the input value to the nearest predefined time range.
+     * 
+     * @param {number} value - The input time value
+     * @returns {number} Snapped time value
+     */
     const snapToNearest = (value) => {
         if (value < 5) return 0;
         if (value <= 15) return 15;
@@ -30,15 +50,31 @@ export const FilterModal = ({ onClose }) => {
         return 60;
     };
 
+    /**
+     * Creates a handler for slider change events that snaps to nearest value.
+     * 
+     * @param {Function} setValue - State setter function
+     * @returns {Function} Event handler for slider changes
+     */
     const handleSliderChange = (setValue) => (e) => {
         const value = Number(e.target.value);
         setValue(snapToNearest(value));
     };
 
+    /**
+     * Handles changes to the number of steps in a recipe.
+     * 
+     * @param {number} steps - Number of recipe steps
+     */
     const handleStepsChange = (steps) => {
         setSteps(steps);
     };
 
+    /**
+     * Handles changes to selected tags, with logging.
+     * 
+     * @param {string[]} tags - Array of selected tags
+     */
     const handleTagsChange = (tags) => {
         console.log("Handling tags change:", tags); // Log the selected tags
         if (JSON.stringify(tags) !== JSON.stringify(selectedTags)) {
@@ -47,6 +83,11 @@ export const FilterModal = ({ onClose }) => {
         }
     };
 
+    /**
+     * Handles changes to selected ingredients, with logging.
+     * 
+     * @param {string[]} ing - Array of selected ingredients
+     */
     const handleIngsChange = (ing) => {
         console.log("Handling ingredients change:", ing); // Log the selected ingredients
         if (JSON.stringify(ing) !== JSON.stringify(selectedIngs)) {
@@ -55,6 +96,10 @@ export const FilterModal = ({ onClose }) => {
         }
     };
 
+    /**
+     * Submits filter criteria by updating URL search parameters.
+     * Resets to the first page and applies selected filters.
+     */
     const handleSubmit = async () => {
         const currentQuery = Object.fromEntries(searchParams.entries());
 
@@ -157,6 +202,14 @@ export const FilterModal = ({ onClose }) => {
     );
 };
 
+/**
+ * Button component to open the filter modal.
+ * 
+ * @component
+ * @param {Object} props - Component props
+ * @param {Function} props.onClick - Callback function to trigger filter modal
+ * @returns {JSX.Element} Rendered filter button
+ */
 export const FilterButton = ({ onClick }) => {
     return (
         <button
@@ -169,10 +222,18 @@ export const FilterButton = ({ onClick }) => {
     );
 };
 
-// Parent Component
+/**
+ * Parent component managing the filter modal's visibility.
+ * 
+ * @component
+ * @returns {JSX.Element} Rendered parent component with filter button and modal
+ */
 const ParentComponent = () => {
     const [isFilterOpen, setIsFilterOpen] = useState(false);
 
+    /**
+     * Toggles the filter modal's open/closed state.
+     */
     const toggleFilterModal = () => {
         setIsFilterOpen((prev) => !prev); // Toggle modal open/close state
     };
